# PIM-7757: Reduce a lot the risk of inserting duplicate completenesses.