--- a/src/Pim/Bundle/CatalogBundle/Doctrine/Common/Saver/ProductModelDescendantsSaver.php
+++ b/src/Pim/Bundle/CatalogBundle/Doctrine/Common/Saver/ProductModelDescendantsSaver.php
@@ -172,8 +172,6 @@ class ProductModelDescendantsSaver implements SaverInterface
      */
     private function computeCompletenesses(array $products): void
     {
-        $this->completenessManager->bulkSchedule($products);
-
         foreach ($products as $product) {
             $this->completenessManager->generateMissingForProduct($product);
             $this->objectManager->persist($product);