--- a/src/Pim/Bundle/CatalogBundle/Doctrine/ORM/CompletenessRemover.php
+++ b/src/Pim/Bundle/CatalogBundle/Doctrine/ORM/CompletenessRemover.php
@@ -69,6 +69,8 @@ class CompletenessRemover implements CompletenessRemoverInterface
 
     /**
      * {@inheritdoc}
+     *
+     * @deprecated Do not use anymore, will be removed in 3.0.
      */
     public function removeForProduct(ProductInterface $product)
     {
@@ -87,6 +89,8 @@ class CompletenessRemover implements CompletenessRemoverInterface
 
     /**
      * {@inheritdoc}
+     *
+     * @deprecated Do not use anymore, will be removed in 3.0.
      */
     public function removeForProductWithoutIndexing(ProductInterface $product): void
     {
@@ -103,6 +107,8 @@ class CompletenessRemover implements CompletenessRemoverInterface
 
     /**
      * {@inheritdoc}
+     *
+     * @deprecated Do not use anymore, will be removed in 3.0.
      */
     public function removeForFamily(FamilyInterface $family)
     {