--- a/src/Pim/Component/Catalog/Completeness/CompletenessRemoverInterface.php
+++ b/src/Pim/Component/Catalog/Completeness/CompletenessRemoverInterface.php
@@ -25,6 +25,8 @@ interface CompletenessRemoverInterface
      * Remove completenesses of a product
      *
      * @param ProductInterface $product
+     *
+     * @deprecated Do not use anymore, will be removed in 3.0.
      */
     public function removeForProduct(ProductInterface $product);
 
@@ -32,6 +34,8 @@ interface CompletenessRemoverInterface
      * Remove completeness of a product without indexing it.
      *
      * @param ProductInterface $product
+     *
+     * @deprecated Do not use anymore, will be removed in 3.0.
      */
     public function removeForProductWithoutIndexing(ProductInterface $product): void;
 
@@ -39,6 +43,8 @@ interface CompletenessRemoverInterface
      * Remove completenesses for all product of a family
      *
      * @param FamilyInterface $family
+     *
+     * @deprecated Do not use anymore, will be removed in 3.0.
      */
     public function removeForFamily(FamilyInterface $family);