… CHANGELOG-2.0.md
… src/Pim/Bundle/CatalogBundle/spec/Doctrine/Common/Saver/ProductModelDescendantsSaverSpec.php
… src/Pim/Bundle/CatalogBundle/spec/Doctrine/Common/Saver/ProductSaverSpec.php
… src/Pim/Component/Catalog/Model/CompletenessInterface.php
… src/Pim/Component/Catalog/spec/Completeness/CompletenessGeneratorSpec.php