## Description

### The issue

We have a lot of concurrent `INSERT INTO` to save the completeness in the database that generates SQL errors because of duplicated keys. For example: Integrity constraint violation: 1062. Duplicate entry 'X-XX-XXXXXX' for key 'searchunique_idx'

Currently, when we save a product, we remove all the completeness in the database, we compute it again and we save its all in the database.

### The solution

In this PR, to avoid the concurrent `INSERT INTO` that could generate a duplicated key, we change the behavior of the completeness computing: we don't remove the existing product completenesses, but we update the existing ones by updating their ratio and we add or delete completenesses if we need.

The downside of this is that we had to make mutable the Completeness object.

### Performance measurement

| Tested with Icecat demo dev | Current 2.0 | With this PR |
|-----------------------------|-------------|--------------|
| Install                     |    180 s    |     183 s    |
| Product import              |     52 s    |      46 s    |

Volumetry is very small, but it seems the fix has no perceptible impacts on performances.

## Definition Of Done

| Q                                 | A
| --------------------------------- | ---
| Added Specs                       | OK
| Added legacy Behats               | -
| Added acceptance tests            | -
| Added integration tests           | -
| Changelog updated                 | OK
| Review and 2 GTM                  | Todo
| Micro Demo to the PO (Story only) | -
| Migration script                  | -
| Tech Doc                          | -

`Todo`: Pending / Work in progress
`OK`: Done / Validated
`-`: Not needed